weather-image-classifier/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-aiplatform==1.*
google-cloud-core==2.*